Add BT71 entry to F1155/F1255
The Modbusmanager from Nibe is missing this entry for the F1255, while it is included in the list of the F1145/F1245. However, you can add this sensor as AUX input in the software in the heatpump.

When tested locally with HASS with Nibe plugin and a F1255-6 running the latest firmware (R7) this sensor register is correctly read-out and displays the same value as on the heatpump.