Add BT74 to coil 47340
If one enables the BT74 sensor as AUX input, the coil 47340 in the F-series does not decode correctly. Upon inspection, it seems that value 18 is sent, which I added to the F-series pumps. Tested locally on a F1255-6 with HA.

I do not know if it actually makes sense to change this value through HA, as the heatpump itself only allows BT74 as an option.