Added Nibe NIBE VVM S330, see https://www.nibe.eu/en-eu/products/heat-pumps/air-water-heat-pumps/vvm-s330, but I need help to complete the PR. I am not very familiar with HA integrations.
For now I added the raw csv in English, created the JSON-file and added it unchanged.
Compared it to S320/S325 and it looks similar in big parts, but is not the same (S330 has a "flow heater" for hot water, while S320 has a drinking water tank).
Apart from the value lists I added the model S330 to the enum.